I was interested in how Russ Cox's nfa handled 0 or 1 alternations not spinning in an infinite loop. Read through his code and wasn't really getting it, so I instrumented it to print out his state graph and then turned that into a graphviz file for visualization.
The way it seems to work is that his implementation will always have the next states list
start with character nodes (in my output circles). And it looks like no chain of diamond
(split) nodes ever loops back on itself. I think he can end up with duplicate states, so
it may be able to create a state explosion if I do it right, will check that next.